Finishing refactoring views.py. Now, all ViewSets are located in the Views directory.
