+++ api/views/__init__.py
@@ -1,2 +1,8 @@
-from .__campus__ import CampusViewSet
-from .__course__ import CourseViewSet
+from .campus_views import CampusViewSet  # noqa: F401
+from .course_views import CourseViewSet  # noqa: F401
+from .subject_views import SubjectViewSet  # noqa: F401
+from .student_views import StudentViewSet  # noqa: F401
+from .emotion_views import EmotionViewSet  # noqa: F401
+from .tag_views import TagViewSet  # noqa: F401
+from .post_views import PostViewSet  # noqa: F401
+from .custom_obtain_jwt_token import CustomObtainJWTToken  # noqa: F401
